changed move_tetros.c

+++ includes/fillit.h
@@ -3,9 +3,6 @@
 
 # define BUFF_SIZE 32
 
-//REMOVE IF SIZE_T IS NEVER USED IN PROTOTYPES
-#include <string.h>
-
 typedef struct			s_tetro
 {
 	unsigned short 		tetro;
@@ -16,7 +13,7 @@ int						get_next_line(const int fd, char **line);
 
 void 					print_tetro(unsigned short tetro);
 unsigned short			convert_tetro(char *s);
-size_t					move_tetros(t_tetro *tetros, size_t tetro_count);
+int						move_tetros(t_tetro *tetros, int tetro_count);
 
 
 #endif
